+++ Untapped/src/main/java/com/skilldistillery/beers/repositories/BeerRepository.java
@@ -15,7 +15,7 @@ public interface BeerRepository extends JpaRepository<Beer, Integer> {
 
 	List<Beer> findByNameContaining(String keyword);
 
-	List<Beer> findByAlcoholBetween(Integer min, Integer max);
+	List<Beer> findByAlcoholBetween(Double min, Double max);
 
 	List<Beer> findByCreatedAtBetween(Date start, Date end);
 }
